Lots of work on getting proper scrolling and track selection

+++ ogl_editor/src/Editor.h
@@ -10,7 +10,7 @@ void Editor_timedUpdate();
 bool Editor_keyDown(int keyCode, int mod);
 void Editor_setWindowSize(int x, int y);
 void Editor_menuEvent(int menuItem);
-void Editor_scroll(int deltaX, int deltaY);
+void Editor_scroll(float deltaX, float deltaY);
 
 enum
 {
